+++ .prettierrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+    singleQuote: false,
+    trailingComma: "all",
+    tabWidth: 4,
+    semi: true,
+    pluginSearchDirs: ["."],
+    svelteSortOrder: "options-styles-scripts-markup",
+};