+++ Changelog.md
@@ -1,7 +1,11 @@
+# v2.0.0
+
+- Update to be compatible with Foundry V13
+
 # v1.1.1
 
-* Fix a bug with pulling players to scenes from the Scene Directory
-* Mark the mod as compatible with V9
+- Fix a bug with pulling players to scenes from the Scene Directory
+- Mark the mod as compatible with V9
 
 # v1.1.0
 
@@ -11,4 +15,4 @@ This version fixes a deprecation warning for 0.8.6+
 
 This is the initial public release of Pull Players to Scene.
 
-* Add a context menu option under scene navigation and in the scenes directory to open a dialog that can be used to select which players to pull to a scene.
+- Add a context menu option under scene navigation and in the scenes directory to open a dialog that can be used to select which players to pull to a scene.